+++ cli/src/main/resources/script_example.txt
@@ -1,8 +1,9 @@
 let a : string = 8 + "th";
 a = readEnv("PATH");
 println(a);
-a = readInput("input string: ");
-println(a);
+let d : number = 3.1;
+d = readInput("input: ");
+println(d);
 let b : bool = true;
