+++ interpreter/src/main/kotlin/CallExpressionInterpreter.kt
@@ -1,3 +1,4 @@
+import inputter.ConsoleInputter
 import outputter.ConsoleOutputter
 
 class CallExpressionInterpreter(
@@ -8,19 +9,8 @@ class CallExpressionInterpreter(
         return when (node.callee.name) {
             "println" -> ConsoleOutputter(variableMap).output(node)
             "readEnv" -> return readEnvironmentVariables(node.arguments[0] as StringLiteral)
-            "readInput" -> return readUserInput(node.arguments[0] as StringLiteral)
+            "readInput" -> return ConsoleInputter(variableMap).readInput(node.arguments[0] as ASTNode)
             else -> throw IllegalArgumentException("Function '${node.callee.name}' not found")
-        }
-    }
-
-    private fun readUserInput(text: StringLiteral?): Any {
-        print(text?.value ?: "")
-        val input = readlnOrNull()
-        return when {
-            input?.matches("true|false".toRegex()) == true -> input.toBoolean()
-            input?.matches("\\d+(\\.\\d+)?".toRegex()) == true -> input.toDoubleOrNull() ?: ""
-            input?.matches("\"[^\"]*\"|'[^']*'".toRegex()) == true -> input.removeSurrounding("\"").removeSurrounding("'")
-            else -> input ?: ""
         }
     }
