+++ interpreter/src/main/kotlin/inputter/ConsoleInputter.kt
@@ -1,0 +1,31 @@
+package inputter
+
+import ASTNode
+import CallExpression
+import CallExpressionInterpreter
+import Identifier
+import IdentifierInterpreter
+import StringLiteral
+import VariableInfo
+
+class ConsoleInputter(private val variableMap: Map<String, VariableInfo>) : Inputter {
+    override fun readInput(node: ASTNode): Any {
+        val text =
+            when (node) {
+                is StringLiteral -> node.value
+                is Identifier -> return IdentifierInterpreter(variableMap).interpret(node)
+                is CallExpression -> return CallExpressionInterpreter(variableMap).interpret(node)
+                else -> ""
+            }
+        print(text)
+        val input = readlnOrNull()
+        return when {
+            input?.matches("true|false".toRegex()) == true -> input.toBoolean()
+            input?.matches("\\d+(\\.\\d+)?".toRegex()) == true -> {
+                input.toIntOrNull() ?: (input.toDoubleOrNull() ?: "")
+            }
+            input?.matches("\"[^\"]*\"|'[^']*'".toRegex()) == true -> input.removeSurrounding("\"").removeSurrounding("'")
+            else -> input ?: ""
+        }
+    }
+}
