+++ interpreter/src/main/kotlin/inputter/Inputter.kt
@@ -1,0 +1,7 @@
+package inputter
+
+import ASTNode
+
+interface Inputter {
+    fun readInput(node: ASTNode): Any
+}
